fixing compiler warnings

+++ wiring/src/spark_wiring_print.cpp
@@ -38,7 +38,7 @@ size_t Print::write(const uint8_t *buffer, size_t size)
 {
   size_t n = 0;
   while (size--) {
-     size_t chunk = write(*buffer++);
+     int chunk = write(*buffer++);
      if (chunk>=0)
          n += chunk;
      else {
